+++ src/main/java/duke/Storage.java
@@ -6,24 +6,39 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
 
+/**
+ * Contains file with specified pathname and related methods.
+ */
 public class Storage {
 
     private final File file;
 
+    /**
+     * Constructor for storage.
+     * @param filePath of stored file.
+     */
     public Storage(String filePath) {
         this.file = new File(filePath);
     }
 
-    public void makeParentDirectories(File file) {
-        File parentFile = file.getParentFile();
+    /**
+     * Creates parent directories if it does not exist.
+     * @param parentFolder parent directory of file that needs to be created.
+     */
+    public void makeParentDirectories(File parentFolder) {
+        File parentFile = parentFolder.getParentFile();
         if (parentFile.exists()) {
-            file.mkdir();
+            parentFolder.mkdir();
         } else {
             makeParentDirectories(parentFile);
-            file.mkdir();
+            parentFolder.mkdir();
         }
     }
 
+    /**
+     * Initialises the file by checking for and creating missing directories and the file.
+     * @return TaskList from reading the file.
+     */
     public ArrayList<Task> setFile() {
         File parentFolder = this.file.getParentFile();
         if (parentFolder.exists()) {
@@ -38,13 +53,17 @@ public class Storage {
         }
     }
 
+    /**
+     * Scans existing file for tasks.
+     * @return ArrayList containing the tasks.
+     */
     public ArrayList<Task> scanExistingFile() {
         ArrayList<Task> tempTaskList = new ArrayList<>();
         try {
             Scanner sc = new Scanner(file);
             while (sc.hasNext()) {
                 String curr = sc.nextLine();
-                Task addTask = Task.fileStringToTask(curr);
+                Task addTask = Parser.fileStringToTask(curr);
                 if (addTask != null) {
                     tempTaskList.add(addTask);
                 }
@@ -55,6 +74,10 @@ public class Storage {
         }
     }
 
+    /**
+     * Creates a new file at file object's file path.
+     * @return Empty ArrayList for TaskList.
+     */
     public ArrayList<Task> createNewFile() {
         try {
             this.file.createNewFile();
